BOAC-1525, separate 'curated' and 'cohort' in front-end /api structure
