+++ src/api/curated.ts
@@ -1,12 +1,6 @@
 import axios from 'axios';
 import store from '@/store';
 
-export function getUsersWithCohorts() {
-  return axios
-    .get(`${store.state.apiBaseUrl}/api/filtered_cohorts/all`)
-    .then(response => response.data, () => null);
-}
-
 export function getCuratedGroup(id) {
   return axios
     .get(`${store.state.apiBaseUrl}/api/curated_cohort/${id}`)
